输出与SDL声道布局一致的数据

+++ ffmpeg_core/ffmpeg_core.h
@@ -75,7 +75,7 @@ FFMPEG_CORE_API int ffmpeg_core_get_bits(MusicHandle* handle);
  * @param handle Handle
  * @return 如果Handle为NULL，返回-1
 */
-FFMPEG_CORE_API int ffmpeg_core_get_bitrate(MusicHandle* handle);
+FFMPEG_CORE_API int64_t ffmpeg_core_get_bitrate(MusicHandle* handle);
 #ifdef __cplusplus
 }
 #endif
